add include typings
Former-commit-id: ce3af5de717ce87a8270d5e9f570314aa1a63be6
Former-commit-id: 9eaccfa2b1188ac7effbca9932f054553aee3670

+++ packages/photon/src/generation/utils.ts
@@ -1,54 +1,72 @@
 import { DMMFClass } from '../runtime/dmmf'
 import { DMMF } from '../runtime/dmmf-types'
 
+export enum Projection {
+  select = 'select',
+  include = 'include',
+}
+
 export function getScalarsName(modelName: string) {
   return `${modelName}Scalars`
 }
 
-export function getPayloadName(modelName: string) {
-  return `${modelName}GetPayload`
+export function getPayloadName(modelName: string, projection: Projection) {
+  return `${modelName}Get${capitalize(projection)}Payload`
 }
+
+// export function getExtractName(modelName: string, projection: Projection) {
+//   return `Extract${modelName}${capitalize(projection)}`
+// }
 
 export function getSelectName(modelName: string) {
   return `${modelName}Select`
+}
+
+export function getIncludeName(modelName: string) {
+  return `${modelName}Include`
 }
 
 export function getDefaultName(modelName: string) {
   return `${modelName}Default`
 }
 
-export function getFieldArgName(field: DMMF.SchemaField): string {
-  return getArgName((field.outputType.type as DMMF.OutputType).name, field.outputType.isList)
+export function getFieldArgName(field: DMMF.SchemaField, projection?: Projection): string {
+  return getArgName((field.outputType.type as DMMF.OutputType).name, field.outputType.isList, projection)
 }
 
-export function getArgName(name: string, isList: boolean): string {
+export function getArgName(name: string, isList: boolean, projection?: Projection): string {
+  const projectionString = projection ? capitalize(projection) : ''
   if (!isList) {
-    return `${name}Args`
+    return `${name}${projectionString}Args`
   }
 
-  return `FindMany${name}Args`
+  return `FindMany${name}${projectionString}Args`
 }
 
 // we need names for all top level args,
 // as GraphQL doesn't have the concept of unnamed args
-export function getModelArgName(modelName: string, action: DMMF.ModelAction): string {
+export function getModelArgName(modelName: string, projection?: Projection, action?: DMMF.ModelAction): string {
+  const projectionName = projection ? capitalize(projection) : ''
+  if (!action) {
+    return `${modelName}${projectionName}Args`
+  }
   switch (action) {
     case DMMF.ModelAction.findMany:
-      return `FindMany${modelName}Args`
+      return `FindMany${modelName}${projectionName}Args`
     case DMMF.ModelAction.findOne:
-      return `FindOne${modelName}Args`
+      return `FindOne${modelName}${projectionName}Args`
     case DMMF.ModelAction.upsert:
-      return `${modelName}UpsertArgs`
+      return `${modelName}${projectionName}UpsertArgs`
     case DMMF.ModelAction.update:
-      return `${modelName}UpdateArgs`
+      return `${modelName}${projectionName}UpdateArgs`
     case DMMF.ModelAction.updateMany:
-      return `${modelName}UpdateManyArgs`
+      return `${modelName}${projectionName}UpdateManyArgs`
     case DMMF.ModelAction.delete:
-      return `${modelName}DeleteArgs`
+      return `${modelName}${projectionName}DeleteArgs`
     case DMMF.ModelAction.create:
-      return `${modelName}CreateArgs`
+      return `${modelName}${projectionName}CreateArgs`
     case DMMF.ModelAction.deleteMany:
-      return `${modelName}DeleteManyArgs`
+      return `${modelName}${projectionName}DeleteManyArgs`
   }
 }
 
@@ -113,6 +131,7 @@ interface SelectReturnTypeOptions {
   hideCondition?: boolean
   isField?: boolean
   fieldName?: string
+  projection: Projection
 }
 
 /**
@@ -130,19 +149,49 @@ export function getSelectReturnType({
 }: SelectReturnTypeOptions) {
   const isList = actionName === DMMF.ModelAction.findMany
 
-  const argName = isField ? getArgName(name, isList) : getModelArgName(name, actionName as DMMF.ModelAction)
+  const selectArgName = isField
+    ? getArgName(name, isList, Projection.select)
+    : getModelArgName(name, Projection.select, actionName as DMMF.ModelAction)
 
+  const includeArgName = isField
+    ? getArgName(name, isList, Projection.include)
+    : getModelArgName(name, Projection.include, actionName as DMMF.ModelAction)
+  
+  const requiredArgName = getModelArgName(name, undefined, actionName as DMMF.ModelAction) + 'Required'
+  const requiredCheck = `T extends ${requiredArgName} ? 'Please either choose \`select\` or \`include\`' : `
+
+  /**
+   * Important: We handle findMany or isList special, as we don't want chaining from there
+   */
   if (isList || hideCondition) {
     const listOpen = isList ? 'Array<' : ''
     const listClose = isList ? '>' : ''
-    return `${renderPromise ? 'Promise<' : ''}${listOpen}${name}GetPayload<Extract${argName}Select<T>>${listClose}${
-      renderPromise ? '>' : ''
-    }`
+    const promiseOpen = renderPromise ? 'Promise<' : ''
+    const promiseClose = renderPromise ? '>' : ''
+    const renderType = (projection: Projection) =>
+      `${promiseOpen}${listOpen}${getPayloadName(name, projection)}<Extract${getModelArgName(
+        name,
+        projection,
+        actionName,
+      )}<T>>${listClose}${promiseClose}`
+
+    return `${requiredCheck}T extends ${selectArgName}
+? ${renderType(Projection.select)} : T extends ${includeArgName}
+? ${renderType(Projection.include)} : ${promiseOpen}${listOpen}${name}${promiseClose}${listClose}`
   }
 
-  return `'select' extends keyof T ? ${renderPromise ? 'Promise<' : ''}${name}GetPayload<Extract${argName}Select<T>>${
-    renderPromise ? '>' : ''
-  } : ${name}Client<${getType(name, isList)}>`
+  const selectType = `${renderPromise ? 'Promise<' : ''}${getPayloadName(
+    name,
+    Projection.select,
+  )}<Extract${selectArgName}<T>>${renderPromise ? '>' : ''}`
+
+  const includeType = `${renderPromise ? 'Promise<' : ''}${getPayloadName(
+    name,
+    Projection.include,
+  )}<Extract${includeArgName}<T>>${renderPromise ? '>' : ''}`
+
+  return `${requiredCheck}T extends ${selectArgName} ? ${selectType}
+: T extends ${includeArgName} ? ${includeType} : ${name}Client<${getType(name, isList)}>`
 }
 
 export function isQueryAction(action: DMMF.ModelAction, operation: 'query' | 'mutation'): boolean {
@@ -151,4 +200,8 @@ export function isQueryAction(action: DMMF.ModelAction, operation: 'query' | 'mu
   }
   const result = action === DMMF.ModelAction.findOne || action === DMMF.ModelAction.findMany
   return operation === 'query' ? result : !result
+}
+
+export function capitalize(str: string) {
+  return str[0].toUpperCase() + str.slice(1)
 }
